pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-optimized-dicom"
version = "0.2.0"
description = "A library for efficiently storing and interacting with DICOM files in the cloud"
readme = "README.md"
authors = [
    {name = "Cal Nightingale", email = "cal@gradienthealth.io"}
]
license = "MIT"
requires-python = ">=3.8"
dependencies = [
    "smart-open==7.0.4",
    "ratarmountcore==0.7.1",
    "numpy",
    "google-cloud-storage==2.19.0",
    "apache-beam[gcp]==2.63.0",
    "filetype==1.2.0",
    "pylibjpeg==2.0.1",
    "pylibjpeg-libjpeg==2.3.0",
    "pylibjpeg-openjpeg==2.4.0",
    "pydicom3>=3.1.0",
    "opencv-python-headless==4.11.0.86",
    "ffmpeg-python==0.2.0",
    "zstandard>=0.24.0",
    # Pin urllib3 to avoid bug in 2.6.x where _GzipDecoder.decompress() doesn't accept max_length parameter
    "urllib3>=2.0.0,<2.6.0",
]

[project.optional-dependencies]
test = [
    "pydicom==2.3.0",
    "matplotlib",
]
dev = [
    "pre-commit>=3.0.0",
    "pydicom==2.3.0",
    "matplotlib",
]

[tool.setuptools]
packages = {find = {}}
include-package-data = true
